requirements.txt
Django
psycopg2
djangorestframework
django-extensions
six
django-cors-middleware
honcho
requests
djangorestframework-jwt
uwsgi
dj-database-url
django-redis
mypy
django-debug-toolbar


flake8
ipython

